update txid scheme & fix test

+++ solgen/src/precompiles/ArbRetryableTx.sol
@@ -61,7 +61,7 @@ interface ArbRetryableTx {
 
     event TicketCreated(bytes32 indexed ticketId);
     event LifetimeExtended(bytes32 indexed ticketId, uint newTimeout);
-    event RedeemScheduled(bytes32 indexed ticketId, bytes32 indexed redeemTxId, uint64 sequenceNum, uint donatedGas, address gasDonor);
+    event RedeemScheduled(bytes32 indexed ticketId, bytes32 indexed redeemTxId, uint64 sequenceNum, uint64 donatedGas, address gasDonor);
     event Redeemed(bytes32 indexed ticketId);
     event Canceled(bytes32 indexed ticketId);
 }
